requirements.txt
pytest~=8.3.5
allure-python-commons~=2.14.0
requests~=2.31.0